refactor: 무한스크롤 콜백함수 setLimit에서 setOffset으로 수정 (#53)
