+++ src/pages/AnswerFeedPage.jsx
@@ -2,23 +2,25 @@ import { useEffect, useState, useRef } from 'react';
 import { useLocation } from 'react-router-dom';
 import { PostHeader, ModalLoading, FeedCardSection } from 'components';
 import { getSubjectsQuestion } from 'api/api';
+import { infiniteScroll } from 'api/infiniteScroll';
 import * as Styled from './StyleFeedPage';
 
-const OFFSET = 0;
+const LIMIT = 1;
 
 const AnswerFeedPage = () => {
   const location = useLocation();
   const subjectId = location.pathname.split('/')[2];
   const target = useRef();
   const [isLoading, setIsLoading] = useState(false);
-  const [limit, setLimit] = useState(1);
+  const [offset, setOffset] = useState(0);
   const [total, setTotal] = useState(null);
   const [subjectName, setSubjectName] = useState('');
   const [subjectImg, setSubjectImg] = useState('');
   const [questionData, setQuestionData] = useState({
     data: [],
   });
-  const option = {filter : true};
+  const option = { filter: true };
+  const observer = infiniteScroll(isLoading, LIMIT, setOffset);
 
   const handleFeedCardSection = async (...args) => {
     setIsLoading(true);
@@ -26,8 +28,7 @@ const AnswerFeedPage = () => {
       const result = await getSubjectsQuestion(...args);
       const { count, results: questionData } = result;
       setQuestionData((prevData) => ({
-        ...prevData,
-        data: questionData,
+        data: [...prevData.data, ...questionData],
       }));
       setTotal(count);
     } catch (error) {
@@ -37,20 +38,9 @@ const AnswerFeedPage = () => {
     }
   };
 
-  const observeOptions = {
-    threshold: 0.6,
-  };
-
-  const observeCallback = () => {
-    if (isLoading) return;
-    setLimit((prev) => prev + 1);
-  };
-
-  const observer = new IntersectionObserver(observeCallback, observeOptions);
-
   useEffect(() => {
-    handleFeedCardSection(subjectId, limit, OFFSET);
-  }, [location, limit]);
+    handleFeedCardSection(subjectId, LIMIT, offset);
+  }, [location, offset]);
 
   useEffect(() => {
     observer.observe(target.current);
